Removing the existing re-import features and texts
