+++ bundles/net.tourbook/src/net/tourbook/Messages.java
@@ -1193,20 +1193,6 @@ public class Messages extends NLS {
    public static String        Import_Data_Action_EditImportPreferences;
    public static String        Import_Data_Action_OpenLogView;
    public static String        Import_Data_Action_OpenLogView_Tooltip;
-   public static String        Import_Data_Action_Reimport_EntireTour;
-   public static String        Import_Data_Action_Reimport_OnlyAltitudeValues;
-   public static String        Import_Data_Action_Reimport_OnlyCadenceValues;
-   public static String        Import_Data_Action_Reimport_OnlyGearValues;
-   public static String        Import_Data_Action_Reimport_OnlyPowerAndPulseValues;
-   public static String        Import_Data_Action_Reimport_OnlyPowerAndSpeedValues;
-   public static String        Import_Data_Action_Reimport_OnlyRunningDynamicsValues;
-   public static String        Import_Data_Action_Reimport_OnlySwimmingValues;
-   public static String        Import_Data_Action_Reimport_OnlyTemperatureValues;
-   public static String        Import_Data_Action_Reimport_OnlyTimeSlices;
-   public static String        Import_Data_Action_Reimport_OnlyTourMarkers;
-   public static String        Import_Data_Action_Reimport_OnlyTourTimerPauses;
-   public static String        Import_Data_Action_Reimport_OnlyTrainingValues;
-   public static String        Import_Data_Action_Reimport_Tour;
    public static String        Import_Data_Action_RemoveTour;
    public static String        Import_Data_Action_RemoveToursWhenClosed;
    public static String        Import_Data_Action_SetupEasyImport_Tooltip;
