+++ bundles/net.tourbook/src/net/tourbook/messages_cs.properties
@@ -1177,19 +1177,6 @@ Import_Data_Action_DeleteTourFiles                               = O&dstranit so
 Import_Data_Action_EditImportPreferences                         = Upravit &p\u0159edvolby importu...
 Import_Data_Action_OpenLogView                                   = Protokol trasy
 Import_Data_Action_OpenLogView_Tooltip                           = Zobraz\u00ED protokol trasy
-Import_Data_Action_Reimport_EntireTour                           = C&elou trasu...
-Import_Data_Action_Reimport_OnlyAltitudeValues                   = Pouze hodnoty n&adm. v\u00FD\u0161ky...
-Import_Data_Action_Reimport_OnlyCadenceValues                    = Pouze hodnoty \u0161lap\u00E1n\u00ED...
-Import_Data_Action_Reimport_OnlyGearValues                       = Pouze hodnoty r&ychlostn\u00EDch stup\u0148\u016F
-Import_Data_Action_Reimport_OnlyPowerAndPulseValues              = Pouze &hodnoty energie a tepu...
-Import_Data_Action_Reimport_OnlyPowerAndSpeedValues              = Pouze hodnoty energie a ryc&hlosti...
-Import_Data_Action_Reimport_OnlyRunningDynamicsValues            = Pouze hodnoty &b\u011B\u017Eeck\u00E9 dynamiky...
-Import_Data_Action_Reimport_OnlySwimmingValues                   = Pouze hodnoty pla&v\u00E1n\u00ED...
-Import_Data_Action_Reimport_OnlyTemperatureValues                = Pouze hodnoty te&ploty...
-Import_Data_Action_Reimport_OnlyTimeSlices                       = V\u0161echny hodno&ty \u010Dasov\u00E9ho \u00FAseku
-Import_Data_Action_Reimport_OnlyTourMarkers                      = Pou&ze zna\u010Dky trasy...
-Import_Data_Action_Reimport_OnlyTrainingValues                   = Pouze tr\u00E9n&inkov\u00FDch hodnot...
-Import_Data_Action_Reimport_Tour                                 = Op\u011Btovn\u00FD import &trasy
 Import_Data_Action_RemoveTour                                    = Odebrat trasu ze seznamu (ulo\u017Een\u00E9 trasy neb&udou smaz\u00E1ny)
 Import_Data_Action_RemoveToursWhenClosed                         = Odeb&rat v\u0161echny trasy z tohoto zobrazen\u00ED, kdy\u017E je zav\u0159eno
 Import_Data_Action_SetupEasyImport_Tooltip                       = Nastaven\u00ED jednoduch\u00E9ho importu.
