+++ bundles/net.tourbook/src/net/tourbook/messages_de.properties
@@ -1193,20 +1193,6 @@ Import_Data_Action_DeleteTourFiles                               = Tour Dateien 
 Import_Data_Action_EditImportPreferences                         = &Import Einstellungen bearbeiten...
 Import_Data_Action_OpenLogView                                   = Tour Protokoll
 Import_Data_Action_OpenLogView_Tooltip                           = Tour Protokoll anzeigen
-Import_Data_Action_Reimport_EntireTour                           = &Gesamte Tour...
-Import_Data_Action_Reimport_OnlyAltitudeValues                   = Nur die &H\u00F6henwerte...
-Import_Data_Action_Reimport_OnlyCadenceValues                    = Nur die Trittfre&quenzwerte...
-Import_Data_Action_Reimport_OnlyGearValues                       = Nur die Werte von Kettenblatt und Rit&zel...
-Import_Data_Action_Reimport_OnlyPowerAndPulseValues              = Nur die Leistungs- und Her&zfrequenzwerte
-Import_Data_Action_Reimport_OnlyPowerAndSpeedValues              = Nur die &Leistungs- und Geschwindigkeitswerte...
-Import_Data_Action_Reimport_OnlyRunningDynamicsValues            = Nur die &Running Dynamics Werte...
-Import_Data_Action_Reimport_OnlySwimmingValues                   = Nur die &Schwimmwerte...
-Import_Data_Action_Reimport_OnlyTemperatureValues                = Nur die &Temperaturwerte...
-Import_Data_Action_Reimport_OnlyTimeSlices                       = Alle &Datenpunkte...
-Import_Data_Action_Reimport_OnlyTourMarkers                      = Nur die Tour&markierungen...
-Import_Data_Action_Reimport_OnlyTourTimerPauses                  = Nur die Tour&pause (Unterbrochene Zeit)...
-Import_Data_Action_Reimport_OnlyTrainingValues                   = Nur die Tra&iningswerte...
-Import_Data_Action_Reimport_Tour                                 = Tour &Reimportieren
 Import_Data_Action_RemoveTour                                    = Tour von Liste &entfernen (Gespeicherte Tour wird nicht gel\u00F6scht)
 Import_Data_Action_RemoveToursWhenClosed                         = Alle Touren von dieser Ansicht &entfernen, wenn sie geschlossen wird
 Import_Data_Action_SetupEasyImport_Tooltip                       = Easy Import konfigurieren.
