+++ bundles/net.tourbook/src/net/tourbook/messages_es.properties
@@ -740,12 +740,6 @@ Image__undo_tour_type_filter_system              = tour-type-filter-system.png
 Image__view_compare_wizard                       = tour-map-compare-wizard.png
 
 Import_Data_Action_EditImportPreferences                    = Editar &Preferencias de Importaci\u00F3n...
-Import_Data_Action_Reimport_EntireTour                      = &Recorrido Completo...
-Import_Data_Action_Reimport_OnlyAltitudeValues              = S\u00F3lo los Valores de &Altitud
-Import_Data_Action_Reimport_OnlyGearValues                  = Solo Valores de Engranajes...
-Import_Data_Action_Reimport_OnlyTemperatureValues           = Solo Valores de Tem&peratura...
-Import_Data_Action_Reimport_OnlyTimeSlices                  = Todos los Valores de Intervalos de Tiempo
-Import_Data_Action_Reimport_Tour                            = &Reimportar Recorrido
 Import_Data_Action_RemoveTour                               = &Eliminar Recorrido de la Lista (Los Recorridos Guardados no son Elimindos)
 Import_Data_Action_RemoveToursWhenClosed                    = &Eliminar todos los recorridos de esta vista al cerrar
 Import_Data_Dialog_ConfirmReimportAltitudeValues_Message    = Al reimportar se reemplazar\u00E1n los valores existentes de altitud, los ajustes de altitud ser\u00E1n perdidos aunque cualquier otra \
