+++ bundles/net.tourbook/src/net/tourbook/messages_fr.properties
@@ -1186,20 +1186,6 @@ Import_Data_Action_DeleteTourFiles                               = &Suppression 
 Import_Data_Action_EditImportPreferences                         = Modifier les &Pr\u00E9f\u00E9rences d'Import
 Import_Data_Action_OpenLogView                                   = Log des parcours
 Import_Data_Action_OpenLogView_Tooltip                           = Afficher le log des parcours
-Import_Data_Action_Reimport_EntireTour                           = &Parcours complet...
-Import_Data_Action_Reimport_OnlyAltitudeValues                   = Seulement les valeurs d'&altitude...
-Import_Data_Action_Reimport_OnlyCadenceValues                    = Seulement les valeurs de &cadence...
-Import_Data_Action_Reimport_OnlyGearValues                       = Seulement les valeurs de &braquet...
-Import_Data_Action_Reimport_OnlyPowerAndPulseValues              = Seulement les valeurs de puissance et fr\u00E9quence cardiaque... 
-Import_Data_Action_Reimport_OnlyPowerAndSpeedValues              = Seulement les valeurs de &puissance et de vitesse...
-Import_Data_Action_Reimport_OnlyRunningDynamicsValues            = Seulement les valeurs de dynamiques de cou&rse...
-Import_Data_Action_Reimport_OnlySwimmingValues                   = Seulement les valeurs de natation...
-Import_Data_Action_Reimport_OnlyTemperatureValues                = Seulement les valeurs de temp\u00E9rature...
-Import_Data_Action_Reimport_OnlyTimeSlices                       = Tous les intervalles de temps
-Import_Data_Action_Reimport_OnlyTourMarkers                      = Seulement les &marqueurs...
-Import_Data_Action_Reimport_OnlyTourTimerPauses                  = Seulement les temps de &pause...
-Import_Data_Action_Reimport_OnlyTrainingValues                   = Seulement les valeurs d'entra\u00EEnement...
-Import_Data_Action_Reimport_Tour                                 = &R\u00E9importer un parcours
 Import_Data_Action_RemoveTour                                    = Reti&rer le parcours de la liste (le parcours enregistr\u00E9 n'est pas supprim\u00E9)
 Import_Data_Action_RemoveToursWhenClosed                         = &Retirer tous les parcours de cette vue quand elle est ferm\u00E9e
 Import_Data_Action_SetupEasyImport_Tooltip                       = Configurer l'assistant d'import.
@@ -3434,7 +3420,7 @@ dialog_reimport_tours_btn_skip_tours_with_importFile_notfound = &Ignorer les par
 dialog_reimport_tours_checkbox_alltours                       = Tous &les tours de la base de donn\u00E9es
 dialog_reimport_tours_checkbox_selectedtours                  = S&eulement les parcours s\u00E9lectionn\u00E9s
 dialog_reimport_tours_dialog_title                            = R\u00E9importer de Parcours
-dialog_reimport_tours_shell_text                              = R\u00E9importer des Parcours...
+dialog_reimport_tours_shell_text                              = &R\u00E9importer des Parcours...
 
 image__merge_tours = merge-tours.png
