+++ bundles/net.tourbook/src/net/tourbook/messages_it.properties
@@ -1167,19 +1167,6 @@ Import_Data_Action_DeleteTourFiles                               = Elimina file 
 Import_Data_Action_EditImportPreferences                         = Modifica &preferenze di importazione...
 Import_Data_Action_OpenLogView                                   = Log percorso
 Import_Data_Action_OpenLogView_Tooltip                           = Mostra il log di percorso
-Import_Data_Action_Reimport_EntireTour                           = Int&ero percorso...
-Import_Data_Action_Reimport_OnlyAltitudeValues                   = Solo i valori di &altitudine...
-Import_Data_Action_Reimport_OnlyCadenceValues                    = Solo i valori di &cadenza...
-Import_Data_Action_Reimport_OnlyGearValues                       = Solo i valori dei rapporti...
-Import_Data_Action_Reimport_OnlyPowerAndPulseValues              = Solo i valori di potenza e di pulsazioni...
-Import_Data_Action_Reimport_OnlyPowerAndSpeedValues              = Solo i valori di potenza e velocit\u00E0...
-Import_Data_Action_Reimport_OnlyRunningDynamicsValues            = Solo i valori di dinamica della cosrsa...
-Import_Data_Action_Reimport_OnlySwimmingValues                   = Solo i valori di nuoto...
-Import_Data_Action_Reimport_OnlyTemperatureValues                = Solo i valori di tem&peratura...
-Import_Data_Action_Reimport_OnlyTimeSlices                       = &Tutti i valori dei campionamenti
-Import_Data_Action_Reimport_OnlyTourMarkers                      = Solo riferimenti percorso...
-Import_Data_Action_Reimport_OnlyTrainingValues                   = Solo i valori di allenamento...
-Import_Data_Action_Reimport_Tour                                 = Reimportare un percorso
 Import_Data_Action_RemoveTour                                    = Rimuo&vi percorso dall'elenco (Il percorso salvato non viene eliminato)
 Import_Data_Action_RemoveToursWhenClosed                         = &Rimuovi tutti i percorsi da questa vista quando si chiude
 Import_Data_Action_SetupEasyImport_Tooltip                       = Imposta l'importazione facilitata.
