+++ bundles/net.tourbook/src/net/tourbook/messages_nl.properties
@@ -1172,19 +1172,6 @@ Import_Data_Action_DeleteTourFiles                               = &Verwijder to
 Import_Data_Action_EditImportPreferences                         = Import&instellingen bewerken...
 Import_Data_Action_OpenLogView                                   = Tocht log
 Import_Data_Action_OpenLogView_Tooltip                           = Toon tocht log
-Import_Data_Action_Reimport_EntireTour                           = &Volledige Tocht...
-Import_Data_Action_Reimport_OnlyAltitudeValues                   = Alleen &hoogte waarden...
-Import_Data_Action_Reimport_OnlyCadenceValues                    = Alleen &cadans waarden
-Import_Data_Action_Reimport_OnlyGearValues                       = Alleen &schakel (wisselen van versnelling) waarden...
-Import_Data_Action_Reimport_OnlyPowerAndPulseValues              = Alleen vermogen- en &hartslagwaarden...
-Import_Data_Action_Reimport_OnlyPowerAndSpeedValues              = Alleen &vermogen- en snelheidswaarden...
-Import_Data_Action_Reimport_OnlyRunningDynamicsValues            = Alleen &hardloopdynamiek waarden...
-Import_Data_Action_Reimport_OnlySwimmingValues                   = Alleen zwemwaarden...
-Import_Data_Action_Reimport_OnlyTemperatureValues                = Alleen temperatuurwaarden...
-Import_Data_Action_Reimport_OnlyTimeSlices                       = Alle &gegevenspunten..
-Import_Data_Action_Reimport_OnlyTourMarkers                      = Alleen tocht markeringen...
-Import_Data_Action_Reimport_OnlyTrainingValues                   = Alleen tra&ining waarden...
-Import_Data_Action_Reimport_Tour                                 = Herimport &tocht
 Import_Data_Action_RemoveTour                                    = Verwijder tocht van lijst (opgeslagen tocht wordt niet verwijderd)
 Import_Data_Action_RemoveToursWhenClosed                         = &Verwijder alle tochten van deze view als het wordt gesloten
 Import_Data_Action_SetupEasyImport_Tooltip                       = Eenvoudige import instellen.
